+++ controllers/management_cluster_network.go
@@ -119,7 +119,7 @@ func (r *ManagementClusterNetworkReconciler) reconcileNormal(ctx context.Context
 func (r *ManagementClusterNetworkReconciler) applyTransitGateway(ctx context.Context, cluster *capa.AWSCluster) error {
 	logger := log.FromContext(ctx)
 
-	id, err := r.transitGateways.Apply(ctx, cluster.Name)
+	id, err := r.transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 	if err != nil {
 		logger.Error(err, "Failed to create transit gateway")
 		return errors.WithStack(err)
@@ -139,7 +139,7 @@ func (r *ManagementClusterNetworkReconciler) applyTransitGateway(ctx context.Con
 func (r *ManagementClusterNetworkReconciler) applyPrefixList(ctx context.Context, cluster *capa.AWSCluster) error {
 	logger := log.FromContext(ctx)
 
-	id, err := r.prefixLists.Apply(ctx, cluster.Name)
+	id, err := r.prefixLists.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 	if err != nil {
 		logger.Error(err, "Failed to create prefix list")
 		return errors.WithStack(err)