+++ tests/integration/aws/transitgateway_test.go
@@ -52,6 +52,9 @@ var _ = Describe("Transitgateway", func() {
 			ObjectMeta: metav1.ObjectMeta{
 				Name: uuid.NewString(),
 			},
+			Spec: capa.AWSClusterSpec{
+				AdditionalTags: additionalTags,
+			},
 		}
 
 		transitGateways, err = awsClients.NewTransitGatewayClient(Region, AwsIamArn)
@@ -60,7 +63,7 @@ var _ = Describe("Transitgateway", func() {
 
 	Describe("Apply", func() {
 		It("creates a transit gateway", func() {
-			arn, err := transitGateways.Apply(ctx, cluster.Name)
+			arn, err := transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 			Expect(err).NotTo(HaveOccurred())
 
 			transitGatewayID, err := aws.GetARNResourceID(arn)
@@ -77,7 +80,7 @@ var _ = Describe("Transitgateway", func() {
 			var originalID string
 
 			BeforeEach(func() {
-				arn, err := transitGateways.Apply(ctx, cluster.Name)
+				arn, err := transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 				Expect(err).NotTo(HaveOccurred())
 
 				originalID, err = aws.GetARNResourceID(arn)
@@ -85,7 +88,7 @@ var _ = Describe("Transitgateway", func() {
 			})
 
 			It("does not return an error", func() {
-				arn, err := transitGateways.Apply(ctx, cluster.Name)
+				arn, err := transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 				Expect(err).NotTo(HaveOccurred())
 
 				actualID, err := aws.GetARNResourceID(arn)
@@ -106,7 +109,7 @@ var _ = Describe("Transitgateway", func() {
 				})
 
 				It("returns an error", func() {
-					_, err := transitGateways.Apply(ctx, cluster.Name)
+					_, err := transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 					Expect(err).To(MatchError(ContainSubstring(
 						"found unexpected number: 2 of transit gatways for cluster",
 					)))