+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add tags from `AWSCluster.Spec.AdditionalTags` and `AWSManagedControlPlane.Spec.AdditionalTags` to  all created resources.
 - Add `global.podSecurityStandards.enforced` value for PSS migration.
 - Add managing prefix lists for management clusters and rename `ManagementClusterTransitGatewayReconciler` to `ManagementClusterNetworkReconciler`.